requirements/requirements-common.txt
opencv-python
numpy
psutil
PyQt6
python-mpv
